+++ src/l1/index.ts
@@ -3,6 +3,7 @@
  *
  * @module
  */
+export * from "./contract-call.js";
 export * from "./dusk-contract-interface.js";
 export * from "./dusk-connect.js";
 export * from "./gas.js";